+++ CHANGELOG.md
@@ -40,7 +40,8 @@ All user visible changes to this project will be documented in this file. This p
     - Emit `TracksApplied` event to create new and update existing tracks ([#105]);
     - `PeerConnection` renegotiation functionality ([#105]);
     - Calculate and send call quality score based on RTC stats ([#132]);
-    - Muting/unmuting `MediaTrack`s by receiver ([#127]).
+    - Muting/unmuting `MediaTrack`s by receiver ([#127]);
+    - Send `TrackUpdate::IceRestart` based on RTC stats analysis ([#138]).
 - [Coturn] integration:
     - [Coturn] sessions destroying ([#84]);
     - [Coturn] stats processing ([#94]).
@@ -79,6 +80,7 @@ All user visible changes to this project will be documented in this file. This p
 [#127]: /../../pull/127
 [#132]: /../../pull/132
 [#135]: /../../pull/135
+[#138]: /../../pull/138